move interval sampler to contrib
